flake::dolphin-emu: correctly set DOLPHIN_WC_REVISION in order to make netplay usable in my configuration
It is not possible to use git describe in a pure system like Nix which hashes all inputs, (since the git index changes whenever anybody pushes a branch, etc), so I have set DOLPHIN_WC_DESCRIBE to the first 8 chars of the scm revision instead of the usual expected value since I have no way of retrieving it. Until https://github.com/dolphin-emu/dolphin/pull/9898 is merged I will be unable to use the Netplay browser, but the scm revision is enough to join Netplay sessions.