examples: init
Initialize examples with an example that shows how to use a webdav mount with an untrusted ssl certificate and username+password, without having to manage state yourself. This is accomplished by making adding to the activation script, though it's not safe and should be managed with sops/agenix, it does show how to avoid managing state imperatively by putting files in /etc or /var/lib.