swordfish: configure plutonomicon hercules-ci-agent via hack
This configures a privileged nixos-container to run hercules-ci-agent. It is passed /run/secrets via a bind mount which is read-only. The group/owner of the secrets is set to 6969, which is also the ID of the user I've made hercules-ci-agent correspond to via users.users.<name>.uid, this is a bit of a hack. Additionally, /nix is writable in the container, due to https://github.com/hercules-ci/hercules-ci-agent/issues/345. Hopefully a fix comes along soon.