swordfish: use hercules-ci multi-agent module
This was causing a lot of issues, running Hercules-CI agents in NixOS
Containers is a bad experience due to systemd-nspawn being immature in
relation to cgroupsv2. Effects would not run in containers and I would
also have to create static UIDs for each instance of the agent and give
those static UIDs to the secrets on the host. This commit uses the
stop-gap module to run multiple agents on the host

https://github.com/hercules-ci/hercules-ci-effects/issues/45

https://github.com/hercules-ci/hercules-ci-agent/pull/359

https://github.com/Plutonomicon/plutonomicon/pull/21