refactor: increase code reuse and implement disks.nix
This commit refactors my config for every machine, it reduces the amount of redundant code and duplication by importing from my profiles/mixins directories. This also means a lot of configuration ends up being identical on every machine, which is good for readability. Additionally, I'm toying with the idea of putting machine-specific configuration in a file that isn't configuration.nix, imagine a file called machine-specific.nix which includes all of the impure real world things such as disk IDs, or even references to WiFi devices that we can't control with Nix.