pyxis: remove prebuilts, enable microg, patch system partition size
Prebuilts cause the system partition to overflow, which means OTA
updates won't work. I have patched this size limit out in the LineageOS
build system, but have also decided that prebuilts are a bad idea
anyway, given that updates from upstream app developers will be
distributed via f-droid. This would only make sense again if we were
able to build the program from source and have it as a flake input.
Microg has recently added Fido support, so I'm enabling it in
preparation for that.